chore(*) do not bundled go-pluginserver with kong builds (#299)
go-pluginserver was included in 2.0 along with Kong distributions. However, as it carries strict requirements of Go compiler version and go-pdk, it ended up not being of much value, since users would be required to recompile it with the same go-pdk and Go compiler as their plugins. It was then decided to remove it, and instead provide guides/examples of how to easily build pluginserver from sources.