feat(*) allow running kong as the 'kong' user (#339)
* feat(*) allow running kong as the 'kong' user

Previously, Kong was, by default, running as root on non-docker systems,
which is not a security best practice. For docker environments, we take
steps to provide more security [2] (create a kong user by default, grant
the same /usr/local/kong/* files and directories ownership to that user
and to the root group) and these are the same steps that are being
employed here for non-docker environments. Post-installation fpm scripts
are being leveraged to create the kong user and change the permissions
of the kong package's contents to that user and to the root group. See
the fpm wiki [1] for details. See all the the files and directories
owned by the kong package here: [files.txt](https://github.com/Kong/kong-build-tools/files/5314643/files.txt)

Note that when `kong start` is run without changing the `$USER`, the
nginx master process will be run as the 'root' user and the nginx
workers processes will run as the 'nobody' user. That is the default
behavior in Kong because the nginx master process needs to run as 'root'
in order for nginx to execute certain actions (listen on privileged
ports such as port 80, for example).

In order to run kong as the 'kong' user, first switch to the kong user:

```
su kong
```

Then start `kong` normally:

```
kong start
```

This will make both the nginx master and worker processes to run as the
'kong' user.

[1] https://github.com/jordansissel/fpm/wiki
[2] https://github.com/Kong/docker-kong/blob/2.1.3/rhel/Dockerfile#L36-L43

* chore(makefile) change DOCKER_KONG_VERSION temporarily

* tests(package) add minimal tests

* tests(package) add chown and chmod tests

* tests(package) validate Kong is running as the correct user

* chore(ci) temporarily remove non-build tests

* tests(package) add 'chmod -R g=u' verification tests

Also fixes a small issue where the user-validation-tests was getting
pulled.

* tests(package) install needed dependencies to run the tests

* tests(package) install debs with dpkg -i /src/kong.deb || apt install --fix-broken -y

Debian jessie can't install local packages with apt-get

* tests(amd64) pin the user-validation tests to as to amd64

* chore(ci) pin DOCKER_KONG_VERSION to master

* Revert "chore(ci) temporarily remove non-build tests"

This reverts commit 66abe497fdbbc414a128c1269cdbc1a9ec35d07a.

Co-authored-by: Colin Hutchinson <chutchic@gmail.com>