fix(openresty-build-tools) remove lua-resty-dns .tar.gz in MacOS
Chaining these actions with `&&` made the `rm -rf` never occur in MacOS.

The `--keep-newer-files` option was making `tar` uncompress the files,
but it was also generating warnings and eventually exiting with a non-0
status.

As a result the .tar.gz file was never removed. When openresty tries to
process lua-resty-dns it got confused by the .tar.gz and would error
with "Multiple entries for lua-resty-dns".