fix(per-commit) add a `nightly` label
Historically, kong-build-tools would update the `latest` tag
for each new per-commit release. Since acfda16, that is no longer the
case. Given it's useful to have a generic name for the latest build of
kong - and we want to avoid pushing a tag named `latest`, so as to avoid
surprises when users pull `kong/kong`- this commit adds a `nightly` tag.
Such tag will be used, for example, in decK's CI, where we want to test
against the latest available build of Kong.