Preparing for 3.0.0 and 3.0.0.0 release
Tested CE and EE release pipelines. The biggest change in this is the
introduction of the KONG_RELEASE_LABEL variable that replaces the
overloaded KONG_VERSION variable. This allows us to label a release with
a version read from meta.lua or a git tag.

Also includes fixes for the FIPS build pipeline, and assorted fixes we
came across in testing.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>