tests(systemd): wait for nginx pidfile
This adds a wait loop that will block until the nginx pidfile is
non-empty (15s max). It should be a little more reliable than the prior
approach of sleeping for 5s and continuing on.

Also, I pinned the worker process count to 2, as it should make for a
consistently faster startup time.