fix(arm64,luajit): bad ptr address with debug mode
We (@dndx ) found that on every Mac M1 aarch64 device (running MacOS or Ubuntu) after building LuaJIT using comand such as 
```
./kong-ngx-build -p buildroot --openresty 1.21.4.1 --openssl 1.1.1s --pcre 8.45 -f --debug
``` the LuaJIT binary cannot init properly, which print

```
./buildroot/openresty/luajit/bin/luajit: cannot create state: not enough memory
```

and @dndx found if `LUAJIT_USE_SYSMALLOC` is enabled in that platform (which is enabled while add `--debug` flag), system malloc function (`realloc` virtually) returns a adress pointer of whom 47 bit is high, but luajit use this bit to express some metadata. So luajit's solution is use mmap to specific such a address that high bits are zero.

So we removed `DLUAJIT_USE_SYSMALLOC` when arch is `aarch64` to make sure luajit get a legal address.